Fix: writing split .dat files are missing total row data

+++ main/logger.h
@@ -120,7 +120,7 @@ typedef struct {
     uint32_t datarows;
     uint8_t numSpiMessages;
     uint32_t msgSize;
-    uint32_t total_datarows;
+    uint64_t total_datarows;
 } sdcard_data_t;
